test: split interception.basic1 into multiple smaller test files to help with parallelization
